batch_pipeline/requirements.txt
apache-airflow==2.7.1
pandas==2.0.3
numpy==1.26.4
